+++ index.js
@@ -6,6 +6,7 @@ export const ALERTS = "/alerts";
 export const MOCKS = "/mocks";
 export const ROUTES = "/routes";
 export const ROUTES_VARIANTS = "/routes-variants";
+export const MOCK_CUSTOM_ROUTES_VARIANTS = "/mock-custom-routes-variants";
 
 export const LEGACY = "/legacy";
 export const BEHAVIORS = "/behaviors";