+++ CHANGELOG.md
@@ -10,6 +10,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 ### Removed
 
+## [2.0.0-beta.2] - 2021-02-14
+
+### Added
+- feat: Add mock-custom-routes-variants path
+
 ## [2.0.0-beta.1] - 2021-02-03
 
 ### Added